Add systemd service
This commit adds a simple systemd service file

To keep it as simple as possible the python script enters
an infinite loop so we can use a systemd
service instead of a cronjob.

The wrapper script used by systemd is used to keep the
systemd service file clean. As we use the
service type 'simple' we can't use multiple PIDs so
the wrapper script becomes the python process with exec